+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "koa": "^2.13.4",
     "koa-bodyparser": "^4.3.0",
+    "koa-controller": "^1.1.0",
     "koa-logger": "^3.2.1",
     "koa-router": "^10.1.1",
     "koa2-cors": "^2.0.6",
