+++ src/controller/user.ts
@@ -1,0 +1,17 @@
+import create from "../service/user/create";
+import { PrismaClient } from "@prisma/client";
+const prisma = new PrismaClient();
+
+export const createUserConstroller = async (data) => {
+    await create(data)
+
+    try {
+       return 'ajout effectuer'
+    }
+    catch(err) {
+        
+    }
+    finally{
+         prisma.$disconnect;
+    };
+}
