--- src/database/database.ts
@@ -1,1 +1,0 @@
-import prismaClient  from '@prisma/client';
