--- src/models/user/read.ts
@@ -1,9 +1,0 @@
-import Router from "koa-router";
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-export const findMany = async () => {
-    const allUsers = await prisma.user.findMany();
-    return allUsers;
-};
